Some tweaks to function declaration syntax
Since I want patterns to be first-class as much as possible,
and because ambiguity arises when the old syntax is used with a
'this'-equivalent argument on the left, I'm introducing a new
pattern-specific syntax. Pattern literals will be surrounded by
vertical bars to set them apart from the rest of the code.